% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{ltrccit}
\alias{ltrccit}
\title{Fit a LBRC conditional inference tree and forest
Fitting LTRC conditional inference tree and forest is also available
We have made used of codes in LTRCforests package, adding some additional functions appropriate for LBRC data}
\usage{
ltrccit(
  formula,
  data,
  id,
  lenbias = FALSE,
  na.action = "na.omit",
  control = partykit::ctree_control()
)
}
\description{
Fit a LBRC conditional inference tree and forest
Fitting LTRC conditional inference tree and forest is also available
We have made used of codes in LTRCforests package, adding some additional functions appropriate for LBRC data
}
