% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictProb.R
\name{predictProb}
\alias{predictProb}
\title{Compute a Survival Curve from ltrrcit and ltrccif model}
\usage{
predictProb(
  object,
  newdata = NULL,
  newdata.id,
  FUN = .pred_Surv_nolog,
  OOB = FALSE,
  time.eval,
  time.tau = NULL
)
}
\description{
Compute a Survival Curve from ltrrcit and ltrccif model
}
