% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune.ltrccif.R
\name{tune.ltrccif}
\alias{tune.ltrccif}
\title{Tune \code{mtry} to the optimal value with respect to out-of-bag error for a ltrccif model}
\usage{
tune.ltrccif(
  formula,
  data,
  id,
  lenbias = FALSE,
  FUN = .pred_Surv_nolog,
  mtryStart = NULL,
  stepFactor = 2,
  time.eval = NULL,
  time.tau = NULL,
  ntreeTry = 100L,
  bootstrap = c("by.sub", "by.root", "none", "by.user"),
  samptype = c("swor", "swr"),
  sampfrac = 0.632,
  samp = NULL,
  na.action = "na.omit",
  trace = TRUE,
  doBest = FALSE,
  plot = FALSE,
  applyfun = NULL,
  cores = NULL,
  control = partykit::ctree_control(teststat = "quad", testtype = "Univ", minsplit =
    nrow(data) * 0.15, minbucket = nrow(data) * 0.06, minprob = 0.01, saveinfo = FALSE)
)
}
\description{
Tune \code{mtry} to the optimal value with respect to out-of-bag error for a ltrccif model
}
